+++ packages/cloud-cognitive/src/components/BreadcrumbWithOverflow/BreadcrumbWithOverflow.stories.js
@@ -27,7 +27,7 @@ export default {
   component: BreadcrumbWithOverflow,
   argTypes: {
     containerWidth: {
-      control: { type: 'range', min: 50, max: 800, step: 10 },
+      control: { type: 'range', min: 20, max: 800, step: 10 },
     },
   },
   decorators: [