+++ packages/cloud-cognitive/src/components/index.js
@@ -8,18 +8,18 @@
 export { AboutModal } from './AboutModal';
 export { ActionBarItem } from './ActionBar';
 export { BreadcrumbWithOverflow } from './BreadcrumbWithOverflow';
-export { EmptyState } from './EmptyStates/EmptyState';
-export { ErrorEmptyState } from './EmptyStates/ErrorEmptyState';
-export { NoDataEmptyState } from './EmptyStates/NoDataEmptyState';
-export { NoTagsEmptyState } from './EmptyStates/NoTagsEmptyState';
-export { NotFoundEmptyState } from './EmptyStates/NotFoundEmptyState';
-export { NotificationsEmptyState } from './EmptyStates/NotificationsEmptyState';
-export { UnauthorizedEmptyState } from './EmptyStates/UnauthorizedEmptyState';
 export { ContextHeader } from './ContextHeader';
+export {
+  EmptyState,
+  ErrorEmptyState,
+  NoDataEmptyState,
+  NoTagsEmptyState,
+  NotFoundEmptyState,
+  NotificationsEmptyState,
+  UnauthorizedEmptyState,
+} from './EmptyStates';
 export { ExampleComponent } from './ExampleComponent';
-export { HTTPError403 } from './HTTPErrors/HTTPError403';
-export { HTTPError404 } from './HTTPErrors/HTTPError404';
-export { HTTPErrorOther } from './HTTPErrors/HTTPErrorOther';
+export { HTTPError403, HTTPError404, HTTPErrorOther } from './HTTPErrors';
 export { ImportModal } from './ImportModal';
 export { ModifiedTabs } from './ModifiedTabs';
 export { Notifications } from './Notifications';